+++ EssentialFeed/EssentialFeediOSTests/Controllers/Helpers/FeedImageCell+TestHelpers.swift
@@ -1,0 +1,40 @@
+//
+//  FeedImageCell+TestHelpers.swift
+//  EssentialFeediOSTests
+//
+//  Created by Alex Tapia on 10/10/21.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension FeedImageCell {
+  
+  func simulateRetryAction() {
+    feedImageRetryButton.simulateTap()
+  }
+  
+  var isShowingLocation: Bool {
+    return !locationContainer.isHidden
+  }
+  
+  var isShowingImageLoadingIndicator: Bool {
+    return feedImageContainer.isShimmering
+  }
+  
+  var isShowingRetryAction: Bool {
+    return !feedImageRetryButton.isHidden
+  }
+  
+  var locationText: String? {
+    return locationLabel.text
+  }
+  
+  var descriptionText: String? {
+    return descriptionLabel.text
+  }
+  
+  var renderedImage: Data? {
+    return feedImageView.image?.pngData()
+  }
+}
